Add "buff_signal(message)" buffer command
Often we want to synchronize animation or sound to the dialogue
appearing in the text boxes. By emitting signals containing messages
at certain points, we can gain this ability.

Using buff_signal(message), we add an object to the buffer with
the message as a parameter. When the _buffer reaches this point,
a signal is emitted with the message as a parameter. The 'message'
can be any object, ie. a string, array or dict - it's up to the
listening object to connect() and handle the messages being emitted.

Example usage:

'''gdscript
tie.buff_text("After hitting ENTER, a signal will be emitted!")
tie.buff_break()
tie.clear()
tie.buff_signal("Player_Animation1")
tie.buff_text("A signal with 'message' of Player_Animation1 has been emitted!")
'''

Adds:
'BUFFER_SIGNAL' - a constant
'custom_message(message)' - a signal with 'message' as a parameter
'buff_signal' - a function